Move UI code from app.js into separate files
