+++ src/ui/ui.js
@@ -1,0 +1,22 @@
+import * as Toolbar from 'ui/toolbar';
+
+function init() {
+    Toolbar.init();
+}
+
+// function updateZoom(scale) {
+//     // 1 / scale * 100
+// }
+
+// function updateMouse(x, y) {}
+
+// function changeTool(tool) {}
+
+// function hoverObject(object) {}
+// function clickObject(object) {}
+// function dragObject(object) {}
+// function dropObject(object) {}
+
+// function updateObject(object) {}
+
+export { init };
